+++ lib/helpers/state_action.dart
@@ -1,30 +1,44 @@
+import 'package:nylo_support/helpers/extensions.dart';
+import 'package:nylo_support/router/router.dart';
+
 import '/alerts/toast_enums.dart';
 import 'helper.dart';
 
 /// [StateAction] class
 class StateAction {
+  /// Helper to find the state name
+  static _findStateName(dynamic state) {
+    if (state is String) {
+      return state;
+    }
+    if (state is RouteView) {
+      return state.nyPageName();
+    }
+    return "";
+  }
+
   /// Refresh the page
-  static refreshPage(String state, {Function()? setState}) {
-    _updateState(state, "refresh-page", {"setState": setState});
+  static refreshPage(dynamic state, {Function()? setState}) {
+    _updateState(_findStateName(state), "refresh-page", {"setState": setState});
   }
 
   /// Set the state of the page
-  static setState(String state, Function() setState) {
-    _updateState(state, "set-state", {"setState": setState});
+  static setState(dynamic state, Function() setState) {
+    _updateState(_findStateName(state), "set-state", {"setState": setState});
   }
 
   /// Pop the page
-  static pop(String state, {dynamic result}) {
-    _updateState(state, "pop", {"setState": result});
+  static pop(dynamic state, {dynamic result}) {
+    _updateState(_findStateName(state), "pop", {"setState": result});
   }
 
   /// Displays a Toast message containing "Sorry" for the title, you
   /// only need to provide a [description].
-  static showToastSorry(String state,
+  static showToastSorry(dynamic state,
       {String? title,
       required String description,
       ToastNotificationStyleType? style}) {
-    _updateState(state, "toast-sorry", {
+    _updateState(_findStateName(state), "toast-sorry", {
       "title": title ?? "Sorry",
       "description": description,
       "style": style ?? ToastNotificationStyleType.danger
@@ -33,11 +47,11 @@ class StateAction {
 
   /// Displays a Toast message containing "Warning" for the title, you
   /// only need to provide a [description].
-  static showToastWarning(String state,
+  static showToastWarning(dynamic state,
       {String? title,
       required String description,
       ToastNotificationStyleType? style}) {
-    _updateState(state, "toast-warning", {
+    _updateState(_findStateName(state), "toast-warning", {
       "title": title ?? "Warning",
       "description": description,
       "style": style ?? ToastNotificationStyleType.warning
@@ -46,11 +60,11 @@ class StateAction {
 
   /// Displays a Toast message containing "Info" for the title, you
   /// only need to provide a [description].
-  static showToastInfo(String state,
+  static showToastInfo(dynamic state,
       {String? title,
       required String description,
       ToastNotificationStyleType? style}) {
-    _updateState(state, "toast-info", {
+    _updateState(_findStateName(state), "toast-info", {
       "title": title ?? "Info",
       "description": description,
       "style": style ?? ToastNotificationStyleType.info
@@ -59,11 +73,11 @@ class StateAction {
 
   /// Displays a Toast message containing "Error" for the title, you
   /// only need to provide a [description].
-  static showToastDanger(String state,
+  static showToastDanger(dynamic state,
       {String? title,
       required String description,
       ToastNotificationStyleType? style}) {
-    _updateState(state, "toast-danger", {
+    _updateState(_findStateName(state), "toast-danger", {
       "title": title ?? "Error",
       "description": description,
       "style": style ?? ToastNotificationStyleType.danger
@@ -72,11 +86,11 @@ class StateAction {
 
   /// Displays a Toast message containing "Oops" for the title, you
   /// only need to provide a [description].
-  static showToastOops(String state,
+  static showToastOops(dynamic state,
       {String? title,
       required String description,
       ToastNotificationStyleType? style}) {
-    _updateState(state, "toast-oops", {
+    _updateState(_findStateName(state), "toast-oops", {
       "title": title ?? "Oops",
       "description": description,
       "style": style ?? ToastNotificationStyleType.danger
@@ -85,11 +99,11 @@ class StateAction {
 
   /// Displays a Toast message containing "Success" for the title, you
   /// only need to provide a [description].
-  static showToastSuccess(String state,
+  static showToastSuccess(dynamic state,
       {String? title,
       required String description,
       ToastNotificationStyleType? style}) {
-    _updateState(state, "toast-success", {
+    _updateState(_findStateName(state), "toast-success", {
       "title": title ?? "Success",
       "description": description,
       "style": style ?? ToastNotificationStyleType.success
@@ -97,11 +111,11 @@ class StateAction {
   }
 
   /// Display a custom Toast message.
-  static showToastCustom(String state,
+  static showToastCustom(dynamic state,
       {String? title,
       required String description,
       ToastNotificationStyleType? style}) {
-    _updateState(state, "toast-custom", {
+    _updateState(_findStateName(state), "toast-custom", {
       "title": title ?? "",
       "description": description,
       "style": style ?? ToastNotificationStyleType.custom
@@ -109,7 +123,7 @@ class StateAction {
   }
 
   /// Validate data from your widget.
-  static validate(String state,
+  static validate(dynamic state,
       {required Map<String, dynamic> rules,
       Map<String, dynamic>? data,
       Map<String, dynamic>? messages,
@@ -120,7 +134,7 @@ class StateAction {
       required Function()? onSuccess,
       Function(Exception exception)? onFailure,
       String? lockRelease}) {
-    _updateState(state, "validate", {
+    _updateState(_findStateName(state), "validate", {
       "rules": rules,
       "data": data,
       "messages": messages,
@@ -134,26 +148,26 @@ class StateAction {
   }
 
   /// Update the language in the application
-  static changeLanguage(String state,
+  static changeLanguage(dynamic state,
       {required String language, bool restartState = true}) {
-    _updateState(state, "change-language", {
+    _updateState(_findStateName(state), "change-language", {
       "language": language,
       "restartState": restartState,
     });
   }
 
   /// Perform a confirm action
-  static confirmAction(String state,
+  static confirmAction(dynamic state,
       {required Function() action,
       required String title,
       String dismissText = "Cancel"}) async {
-    _updateState(state, "confirm-action",
+    _updateState(_findStateName(state), "confirm-action",
         {"action": action, "title": title, "dismissText": dismissText});
   }
 
   /// Updates the page [state]
   /// Provide an [action] and [data] to call a method in the [NyState].
-  static void _updateState(String state, String action, dynamic data) {
+  static void _updateState(dynamic state, String action, dynamic data) {
     updateState(state, data: {"action": action, "data": data});
   }
 }